refactor: import moduls core layer. Modify imports in affected layers
