+++ adapters/src/repositories/memory/memory_appointment_repository.py
@@ -1,7 +1,7 @@
 from typing import List, Optional
 
-from core.src.models.appointment import Appointment
-from core.src.repository.appointment_repository import AppointmentRepository
+from core.src.models import Appointment
+from core.src.repository import AppointmentRepository
 
 
 class MemoryAppointmentRepository(AppointmentRepository):
