+++ api/src/routers/product.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, HTTPException
 
 from api.src.dtos import Product
-from core.src.use_cases.product.create.request import CreateProductRequest
+from core.src.use_cases.product import CreateProductRequest
 from factories.use_cases.product import (create_product_use_case,
                                          get_all_products_use_case)
