+++ core/src/exceptions/business/__init__.py
@@ -1,0 +1,3 @@
+from .appointment import (AppointmentAlreadyExistsException,
+                          AppointmentBusinessException)
+from .product import ProductAlreadyExistsException, ProductBusinessException
