+++ core/src/models/__init__.py
@@ -1,0 +1,2 @@
+from .appointment import Appointment
+from .product import Product
