+++ core/src/use_cases/__init__.py
@@ -1,0 +1,6 @@
+from .appointment import (CreateAppointment, CreateAppointmentRequest,
+                          CreateAppointmentResponse, GetAllAppointments,
+                          GetAllAppointmentsResponse)
+from .product import (CreateProduct, CreateProductRequest,
+                      CreateProductResponse, GetAllProducts,
+                      GetAllProductsResponse)
