+++ core/src/use_cases/appointment/__init__.py
@@ -1,2 +1,3 @@
-from .create import CreateAppointment
-from .get_all import GetAllAppointments
+from .create import (CreateAppointment, CreateAppointmentRequest,
+                     CreateAppointmentResponse)
+from .get_all import GetAllAppointments, GetAllAppointmentsResponse
