+++ core/src/use_cases/product/__init__.py
@@ -1,2 +1,2 @@
-from .create import CreateProduct
-from .get_all import GetAllProducts
+from .create import CreateProduct, CreateProductRequest, CreateProductResponse
+from .get_all import GetAllProducts, GetAllProductsResponse
