+++ core/tests/fixtures/product_fixtures.py
@@ -1,10 +1,8 @@
 import pytest
 
-from adapters.src.repositories.memory.memory_product_repository import \
-    MemoryProductRepository
-from core.src.use_cases.product.create import (CreateProduct,
-                                               CreateProductRequest)
-from core.src.use_cases.product.get_all.use_case import GetAllProducts
+from adapters.src.repositories import MemoryProductRepository
+from core.src.use_cases import (CreateProduct, CreateProductRequest,
+                                GetAllProducts)
 
 
 @pytest.fixture
