+++ factories/repositories/__init__.py
@@ -1,0 +1,3 @@
+from .appointment import sql_appointment_repository
+from .email_sender import email_sender_repository
+from .product import sql_product_repository
