+++ factories/repositories/email_sender.py
@@ -1,6 +1,5 @@
-from adapters.src.repositories.email_sender.email_sender_adapter import \
-    SMTPEmailSenderRepository
-from core.src.repository.email_sender_repository import EmailSenderRepository
+from adapters.src.repositories import SMTPEmailSenderRepository
+from core.src.repository import EmailSenderRepository
 
 
 def email_sender_repository() -> EmailSenderRepository:
