+++ factories/use_cases/__init__.py
@@ -1,0 +1,2 @@
+from .appointment import sql_appointment_repository
+from .product import sql_product_repository
